Make product-related tests granular
