+++ plugins/woocommerce/changelog/e2e-add-new-test-shopper-product
@@ -1,4 +1,4 @@
 Significance: patch
 Type: dev
 
-Add a new E2E test covering shopper product page
+Add new E2E test covering shopper product page and make Product-related tests granular (separated test files)
